requirements.txt
PySide6>=6.6.1
cryptography>=41.0.0
base58>=2.1.1
pybase64>=1.3.0
python-magic>=0.4.27
qrcode>=7.4.2
pillow>=10.2.0
pycryptodome>=3.19.0
